create abstract class for streaming dataset
